+++ app/src/main/java/com/example/extrack/AdapterExpense.java
@@ -33,7 +33,7 @@ public class AdapterExpense extends RecyclerView.Adapter<AdapterExpense.MyViewHo
     private Button btnUpdate,btnDelete;
 
     private String type,note,post_key;
-    private int amount;
+    private float amount;
 
     private DatabaseReference mIncomeData;
 
@@ -112,7 +112,7 @@ public class AdapterExpense extends RecyclerView.Adapter<AdapterExpense.MyViewHo
                 note=editNote.getText().toString().trim();
                 String maomunt=String.valueOf(amount);
                 maomunt=editAmount.getText().toString().trim();
-                int myAmount=Integer.parseInt(maomunt);
+                float myAmount=Float.parseFloat(maomunt);
                 String mDate = DateFormat.getDateInstance().format(new Date());
                 Data data = new Data(myAmount,type,note,post_key,mDate);
